start cleanup kernel
