+++ tests/tt_metal/tt_metal/api/test_runtime_args.cpp
@@ -37,6 +37,7 @@
 
 // Access to internal API: ProgramImpl::num_kernel, get_kernel
 #include "impl/program/program_impl.hpp"
+#include "impl/kernels/kernel_impl.hpp"
 
 using namespace tt;
 using namespace tt::tt_metal;
