+++ tests/tt_metal/tt_metal/test_compile_program.cpp
@@ -40,6 +40,7 @@
 
 // Access to internal API: ProgramImpl::num_kernel, get_kernel
 #include "impl/program/program_impl.hpp"
+#include "impl/kernels/kernel_impl.hpp"
 
 using std::vector;
 using namespace tt;
