+++ tt_metal/fabric/fabric.cpp
@@ -22,6 +22,7 @@
 
 #include "impl/context/metal_context.hpp"
 #include "impl/program/program_impl.hpp"
+#include "impl/kernels/kernel_impl.hpp"
 #include <umd/device/types/xy_pair.hpp>
 
 #include "fabric_host_utils.hpp"
