+++ tt_metal/impl/debug/inspector.cpp
@@ -12,6 +12,7 @@
 #include "program.hpp"
 #include <memory>
 #include <tt-logger/tt-logger.hpp>
+#include "impl/kernels/kernel_impl.hpp"
 
 namespace tt::tt_metal {
