+++ tt_metal/impl/dispatch/data_collection.cpp
@@ -23,6 +23,7 @@
 #include "assert.hpp"
 #include "hal_types.hpp"
 #include "impl/context/metal_context.hpp"
+#include "impl/kernels/kernel_impl.hpp"
 #include "program/program_impl.hpp"
 #include "tt-metalium/program.hpp"
